service skel

+++ core/model/link.h
@@ -5,9 +5,9 @@
 
 class Link {
 public:
-	Link(unsigned int id, Node node1, Node node2) : id(id), node1(node1), node2(node2) {};
+	Link(unsigned long id, Node node1, Node node2) : id(id), node1(node1), node2(node2) {};
 private:
-	unsigned int id;
+	unsigned long id;
 	Node node1;
 	Node node2;
 };
