+++ core/model/user.h
@@ -7,9 +7,9 @@
 
 class User {
 public:
-    User(unsigned int id, std::string name, std::string password_hash, UserRole role) : id(id), name(name), password_hash(password_hash), role(role) {};
+    User(unsigned long id, std::string name, std::string password_hash, UserRole role) : id(id), name(name), password_hash(password_hash), role(role) {};
 private:
-    unsigned int id;
+    unsigned long id;
     std::string name;
     std::string password_hash;
     UserRole role;
